[Praveen] Add: Implemented the functionality to edit the details of t…

+++ js/EmployeePayrollHome.js
@@ -5,9 +5,14 @@ window.addEventListener("DOMContentLoaded", (event) => {
   /// Calling the get employee payroll data from storage method to populate the empPayrollList
   empPayrollList = getEmployeePayrollDataFromStorage();
   /// Selecting the emp-count class element to update the previously used static count of elements
-  document.querySelector(".emp-count").textContent = empPayrollList.length;
+  var element = document.querySelector(".emp-count");
+  if (element) {
+    element.textContent = empPayrollList.length;
+  }
   /// Calling the inner HTML method to initialise a header template and then bind it with the table
   createInnerHtml();
+  /// Removing data from the emp edit
+  localStorage.removeItem("editEmp");
 });
 /// UC4 -- Using the template literal feature of ES6
 const createInnerHtml = () => {
@@ -16,7 +21,7 @@ const createInnerHtml = () => {
   /// Defining the common header html tags for the table to be displayed at the page
   const headerHtml =
     "<th></th><th>Name</th><th>Gender</th><th>Department</th><th>Salary</th><th>Start Date</th><th>Actions</th>";
-    /// Defining the inner html tag in which all the elements will be appended to populate the table
+  /// Defining the inner html tag in which all the elements will be appended to populate the table
   let innerHtml = `${headerHtml}`;
   /// Iterating over the list of elements in employee payroll list
   /// Continuously appending each row structure of html page to the innerHtml tag using placeholder, template literals
@@ -36,7 +41,7 @@ const createInnerHtml = () => {
             empPayrollData._id
           }" onclick= "remove(this)" alt="delete" src="../assets/icons/delete-black-18dp.svg">
             <img id="${
-              empPayrollData.id
+              empPayrollData._id
             }" onclick= "update(this)" alt="edit" src="../assets/icons/create-black-18dp.svg"></td>
     </tr>`;
   }
@@ -87,19 +92,38 @@ const getDeptHtml = (deptList) => {
 };
 
 /// UC1 -- Remove the employee data from the local storage
- const remove= (node)=>{
+const remove = (node) => {
   /// Finding whether the data is present in the employee payroll list or not
-  let empPayrollData= empPayrollList.find(empData=>empData._id=node.id);
-  /// Finding whether the element is found or not in the local storage of the employee salary list
-  if(!empPayrollData) return;
-  /// Finding the index position of the employee payroll data in the employee payroll list using map and index method
-  const index= empPayrollList.map(empData=>empData._id).indexOf(empPayrollData.id);
-  /// Deleting the found data from the employee salary list(Not from the local storage till)
-  empPayrollList.splice(index-1,1);
-  /// Pushing the updated data to the local storage- updation => deleting the passed node
-  localStorage.setItem("EmployeePayrollList",JSON.stringify(empPayrollList));
-  /// Count is updated on the web page so as to see the changed count
-  document.querySelector(".emp-count").textContent= empPayrollList.length;
+  let empPayrollData = empPayrollList.find(
+    (empData) => (empData._id = node.id)
+  );
+  var result = confirm("Want to delete the data for Employee : " + empPayrollData._name);
+  if (result) {
+    /// Finding whether the element is found or not in the local storage of the employee salary list
+    if (!empPayrollData) return;
+    /// Finding the index position of the employee payroll data in the employee payroll list using map and index method
+    const index = empPayrollList
+      .map((empData) => empData._id)
+      .indexOf(empPayrollData.id);
+    /// Deleting the found data from the employee salary list(Not from the local storage till)
+    empPayrollList.splice(index - 1, 1);
+    /// Pushing the updated data to the local storage- updation => deleting the passed node
+    localStorage.setItem("EmployeePayrollList", JSON.stringify(empPayrollList));
+    /// Count is updated on the web page so as to see the changed count
+    document.querySelector(".emp-count").textContent = empPayrollList.length;
+  }
   /// Using the create inner html to display the tabular data on the home webpage
   createInnerHtml();
-} 
+};
+
+/// UC2 -- To edit detail defining the update method with parameter as the node
+const update = (node) => {
+  /// Finding whether the data is present in the employee payroll list or not
+  let empPayrollData = empPayrollList.find((empData) => empData._id == node.id);
+  /// Finding whether the element is found or not in the local storage of the employee salary list
+  if (!empPayrollData) return;
+  /// Creating a different list for employee payroll data and storing it in the local storage
+  localStorage.setItem("editEmp", JSON.stringify(empPayrollData));
+  /// Redirecting to thee employee payroll page
+  window.location.replace(site_properties.emp_payroll_page);
+};
